Delete the rest of v1 `pants.backend.python`
# Rust tests will be skipped. Delete if not intended.
[ci skip-rust-tests]
