+++ src/python/pants/backend/project_info/BUILD
@@ -28,7 +28,6 @@ python_tests(
     ':project_info',
     'src/python/pants/backend/codegen/protobuf',
     'src/python/pants/backend/python/rules',
-    'src/python/pants/backend/python/targets',
     'src/python/pants/core/util_rules',
     'src/python/pants/engine:addresses',
     'src/python/pants/engine:fs',
