+++ src/python/pants/backend/python/BUILD
@@ -11,7 +11,6 @@ python_library(
     ':target_types',
     'src/python/pants/backend/python/dependency_inference',
     'src/python/pants/backend/python/rules',
-    'src/python/pants/backend/python/targets',
     'src/python/pants/build_graph',
     'src/python/pants/goal:task_registrar',
     'src/python/pants/python',
@@ -23,7 +22,6 @@ python_library(
   name = 'all_utils',
   sources = ['*.py', '!__init__.py', '!register.py', '!*_test.py'],
   dependencies = [
-    ':interpreter_cache',
     ':python_artifact',
     ':python_requirements',
     'src/python/pants/python',
@@ -32,26 +30,9 @@ python_library(
 )
 
 python_library(
-  name = 'interpreter_cache',
-  sources = ['interpreter_cache.py'],
-  dependencies = [
-    '3rdparty/python:pex',
-    'src/python/pants/backend/python/targets',
-    'src/python/pants/base:exceptions',
-    'src/python/pants/process',
-    'src/python/pants/python',
-    'src/python/pants/subsystem',
-    'src/python/pants/util:dirutil',
-    'src/python/pants/util:memo',
-  ],
-  tags = {"partially_type_checked"},
-)
-
-python_library(
   name = 'pants_requirement',
   sources = ['pants_requirement.py'],
   dependencies = [
-    'src/python/pants/backend/python/targets',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:exceptions',
     'src/python/pants/build_graph',
@@ -82,7 +63,6 @@ python_library(
   dependencies = [
     ':python_artifact',
     'src/python/pants/backend/python/subsystems',
-    'src/python/pants/backend/python/targets',
     'src/python/pants/core',
     'src/python/pants/core/util_rules',
     'src/python/pants/engine:addresses',
