--- src/python/pants/backend/python/targets/BUILD
@@ -1,17 +1,0 @@
-# Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
-# Licensed under the Apache License, Version 2.0 (see LICENSE).
-
-python_library(
-  dependencies = [
-    '3rdparty/python:pex',
-    'src/python/pants/backend/python:python_artifact',
-    'src/python/pants/base:exceptions',
-    'src/python/pants/base:payload',
-    'src/python/pants/base:payload_field',
-    'src/python/pants/build_graph',
-    'src/python/pants/python',
-    'src/python/pants/util:collections',
-    'src/python/pants/util:ordered_set',
-  ],
-  tags = {"partially_type_checked"},
-)
