+++ src/python/pants/init/engine_initializer.py
@@ -5,9 +5,6 @@ import logging
 from dataclasses import dataclass
 from typing import Any, Iterable, List, Optional, Set, Tuple, Type, cast
 
-from pants.backend.python.targets.python_binary import PythonBinary
-from pants.backend.python.targets.python_library import PythonLibrary
-from pants.backend.python.targets.python_tests import PythonTests
 from pants.base.build_environment import get_buildroot
 from pants.base.build_root import BuildRoot
 from pants.base.exiter import PANTS_SUCCEEDED_EXIT_CODE
@@ -32,13 +29,7 @@ from pants.engine.internals.scheduler import Scheduler, SchedulerSession
 from pants.engine.legacy.address_mapper import LegacyAddressMapper
 from pants.engine.legacy.graph import LegacyBuildGraph, create_legacy_graph_tasks
 from pants.engine.legacy.parser import LegacyPythonCallbacksParser
-from pants.engine.legacy.structs import (
-    PythonBinaryAdaptor,
-    PythonTargetAdaptor,
-    PythonTestsAdaptor,
-    RemoteSourcesAdaptor,
-    TargetAdaptor,
-)
+from pants.engine.legacy.structs import RemoteSourcesAdaptor, TargetAdaptor
 from pants.engine.legacy.structs import rules as structs_rules
 from pants.engine.platform import create_platform_rules
 from pants.engine.rules import RootRule, rule
@@ -94,9 +85,6 @@ def _apply_default_sources_globs(
 # TODO: These calls mutate the adaptor classes for some known library types to copy over
 # their default source globs while preserving their concrete types. As with the alias replacement
 # below, this is a delaying tactic to avoid elevating the TargetAdaptor API.
-_apply_default_sources_globs(PythonBinaryAdaptor, PythonBinary)
-_apply_default_sources_globs(PythonTargetAdaptor, PythonLibrary)
-_apply_default_sources_globs(PythonTestsAdaptor, PythonTests)
 _apply_default_sources_globs(RemoteSourcesAdaptor, RemoteSources)
 
 
@@ -142,9 +130,6 @@ def _legacy_symbol_table(
         }
     )
 
-    table["python_library"] = PythonTargetAdaptor
-    table["python_tests"] = PythonTestsAdaptor
-    table["python_binary"] = PythonBinaryAdaptor
     table["remote_sources"] = RemoteSourcesAdaptor
 
     return SymbolTable(table)
