+++ tests/python/pants_test/backend/python/BUILD
@@ -6,7 +6,6 @@ python_tests(
   sources=['test_pants_requirement.py'],
   dependencies=[
     'src/python/pants/backend/python:plugin',
-    'src/python/pants/backend/python/targets',
     'src/python/pants/base:build_environment',
     'src/python/pants/python',
     'src/python/pants/testutil:test_base',
@@ -18,25 +17,10 @@ python_tests(
   name='python_requirement_list',
   sources=['test_python_requirement_list.py'],
   dependencies=[
-    'src/python/pants/backend/python/targets',
     'src/python/pants/base:exceptions',
     'src/python/pants/build_graph',
     'src/python/pants/python',
     'src/python/pants/testutil:test_base'
-  ],
-  tags = {"partially_type_checked"},
-)
-
-python_tests(
-  name = 'interpreter_cache',
-  sources = ['test_interpreter_cache.py'],
-  dependencies = [
-    'src/python/pants/backend/python/subsystems',
-    'src/python/pants/backend/python:interpreter_cache',
-    'src/python/pants/util:contextutil',
-    'src/python/pants/testutil:pexrc_util',
-    'src/python/pants/testutil:interpreter_selection_utils',
-    'src/python/pants/testutil:test_base',
   ],
   tags = {"partially_type_checked"},
 )
